Added error codes, moved the attls actions to bin/libs and updated references
