+++ bin/commands/.errors
@@ -56,3 +56,7 @@ ZWEL0322E|322|%s is not a valid directory.
 ZWEL0323E|323|Certificate validation failed. Fix errors listed before starting Zowe.
 ZWEL0326E|326|An error occurred while processing Zowe YAML config %s:
 ZWEL0327E|327|Failed to read %s - %s.
+ZWEL0363W||%s is not %s. Skipping port ATTLS validation. 
+ZWEL0364E|364|%s is not %s. Skipping port ATTLS validation.
+ZWEL0365E|365|%s: No AT-TLS rule identified on %s:%d for user %s and jobname %s
+ZWEL0366E|366|%d Zowe port ATTLS validation(s) failed, review output for action items before running Zowe.
